MDM expansion doesn't work when using the :mdm qualifier in query string (#5803)
* wip

* cleaning up test

* adding changelog and passing spotless.

---------
